pyproject.toml
[tool.poetry]
name = "yamlscript"
version = "0.1.0"
description = ""
authors = ["Stanislaw Babkin <stalissss@outlook.com>"]
readme = "README.md"

[tool.poetry.scripts]
yamlscript = "yamlscript.main:main"

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.4.2"
pyyaml = "^6.0.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
zimports = "^0.6.1"
pre-commit = "^3.4.0"

[tool.poetry-dynamic-versioning]
enable = true

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"
